npi de que se hizo, debido a que compila con errores al crear la instancia, se hizo una copia en Visual studio 2019 y ahi si compila y debido a que la herramienta del debugger aqui en code no quiere funcionar se migra a Visual Studio 2019

+++ main.cpp
@@ -32,7 +32,7 @@ void menu_ordenar()
 int main()
 {
      setlocale(LC_ALL, "spanish");
-     int opcion, cantidad;
+     size_t opcion, cantidad;
      cout << "Ingrese la cantidad de Jugadores: ";
      cin >> cantidad;
      VideoJuego vj(cantidad);
